Render strategy pattern
